requirements.txt
    openai
    python-dotenv
    langchain-openai	
    langchain_google_genai 
    langchain-ollama
    langchain_core
    langchain_ollama
    streamlit
    gradio
    langchain-huggingface
    huggingface_hub
    transformers
    torch
    langchain
    langchain_community
    langchain-aws	
    groq
    langchain_groq
    pypdf
    bs4